+++ src/main/java/qna/service/QnaService.java
@@ -4,9 +4,10 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qna.CannotDeleteException;
-import qna.NotFoundException;
+
 import qna.domain.*;
+import qna.exception.CannotDeleteException;
+import qna.exception.NotFoundException;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
